Rename files, move tests
