+++ common/persistence/visibility/client/visibility_manager.go
