+++ common/persistence/visibility/visibility_manager_dual_write.go
